MINOR: tinfo/clock: turn sched_call_date to 64-bits
We used to store it in 32-bits since we'd only use it for latency and CPU
usage calculation but usages will evolve so let's not truncate the value
anymore. Now we store the full 64 bits. Note that this doesn't even
increase the storage size due to alignment. The 3 usage places were
verified to still be valid (most were already cast to 32 bits anyway).

+++ include/haproxy/tinfo-t.h
@@ -154,9 +154,7 @@ struct thread_ctx {
 	uint emergency_bufs_left;           /* number of emergency buffers left in magic_bufs[] */
 
 	uint32_t sched_wake_date;           /* current task/tasklet's wake date in 32-bit ns or 0 if not supported */
-	uint32_t sched_call_date;           /* current task/tasklet's call date in 32-bit ns */
-
-	// 4 bytes hole here
+	uint64_t sched_call_date;           /* current task/tasklet's call date in ns */
 
 	uint64_t prev_mono_time;            /* previous system wide monotonic time (leaving poll) */
 	uint64_t curr_mono_time;            /* latest system wide monotonic time (leaving poll) */
